hooked grass functions into code
